Remove Image/AbstractJobTest
